factored places api endpoints
